Convert money/quantity/weight columns to Numeric and fix Decimal arithmetic
Migration 024 converts Float money/quantity/weight DB columns to Numeric
for exact fixed-point arithmetic (Pydantic schemas now validate as Decimal
via app/schemas/money.py). This surfaced TypeError crashes wherever service
code mixed float and Decimal: journal posting/reporting in
accounting_service.py, InventoryCoordinator's average-cost calculation, and
sales/slaughter/feed/branch-transfer flows that pass Decimal quantities into
still-float inventory APIs.

Add app/core/money.py (to_decimal/quantize_money, ROUND_HALF_UP) and use it
throughout accounting_service.py and inventory_coordinator.py; cast Decimal
quantities to float at StockMovement/coordinator boundaries; fix
InventoryCoordinator.record_movement missing batch_id parameter. Add
regression tests covering Decimal-safe ledger/trial-balance/P&L/balance-sheet/
cash-flow reporting and average-cost calculation.

+++ backend/app/models/branch_transfer.py
@@ -8,7 +8,7 @@ Transfers go into a virtual 'transit' state until received by the destination br
 from datetime import datetime, timezone
 import enum
 
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, String, Text, UniqueConstraint
+from sqlalchemy import Column, DateTime, Numeric, ForeignKey, Integer, String, Text, UniqueConstraint
 from sqlalchemy.orm import relationship
 
 from app.db.base import Base
@@ -63,8 +63,8 @@ class BranchTransferItem(Base):
     transfer_id = Column(Integer, ForeignKey("branch_transfers.id", ondelete="CASCADE"), nullable=False, index=True)
     stock_item_id = Column(Integer, ForeignKey("stock_items.id"), nullable=False)
     
-    quantity_shipped = Column(Float, nullable=False)
-    quantity_received = Column(Float, nullable=True)
+    quantity_shipped = Column(Numeric(14, 4), nullable=False)
+    quantity_received = Column(Numeric(14, 4), nullable=True)
     
     notes = Column(String(255), nullable=True)
